+++ bindings/actuator_ffi.h
@@ -4,11 +4,11 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-typedef float deflection_t;
+typedef double deflection_t;
 
 typedef struct {
     deflection_t target_deflection;
-    float rate_limit;
+    double rate_limit;
 } command_t;
 
 typedef struct {
@@ -22,11 +22,11 @@ typedef enum {
     FAILED = 2
 } health_status_t;
 
-bool validate_command(float current, float target, float max_rate, float max_def, float min_def);
+bool validate_command(double current, double target, double max_rate, double max_def, double min_def);
 command_t apply_limits(command_t cmd, limits_t limits);
-bool is_within_limits(float value, limits_t limits);
+bool is_within_limits(double value, limits_t limits);
 health_status_t check_overall_status(uint8_t pos_status, uint8_t temp_status, uint8_t curr_status, uint8_t hyd_status);
 command_t majority_vote(command_t c1, command_t c2, command_t c3, limits_t limits);
-bool is_consensus(command_t c1, command_t c2, command_t c3, float tolerance);
+bool is_consensus(command_t c1, command_t c2, command_t c3, double tolerance);
 
 #endif
