+++ bindings/ocaml_glue.h
@@ -5,20 +5,20 @@
 #include <stddef.h>
 
 typedef struct {
-    float position;
-    float velocity;
+    double position;
+    double velocity;
 } setpoint_t;
 
 typedef struct {
     setpoint_t* points;
     size_t count;
-    float duration;
-    float max_rate;
+    double duration;
+    double max_rate;
 } trajectory_data_t;
 
 typedef struct {
-    float target_deflection;
-    float rate_limit;
+    double target_deflection;
+    double rate_limit;
 } actuator_cmd_t;
 
 typedef enum {
@@ -27,15 +27,15 @@ typedef enum {
     GLUE_EMERGENCY = 2
 } glue_flight_mode_t;
 
-actuator_cmd_t glue_send_trajectory_point(setpoint_t point, float max_rate);
+actuator_cmd_t glue_send_trajectory_point(setpoint_t point, double max_rate);
 actuator_cmd_t glue_finalize_trajectory(trajectory_data_t trajectory);
 void glue_abort_trajectory(void);
 bool glue_is_trajectory_active(void);
 glue_flight_mode_t glue_get_current_mode(void);
 void glue_set_mode(glue_flight_mode_t mode);
-actuator_cmd_t glue_get_emergency_command(float deflection);
-float glue_get_min_deflection(void);
-float glue_get_max_deflection(void);
-bool glue_validate_external_command(float deflection, float rate);
+actuator_cmd_t glue_get_emergency_command(double deflection);
+double glue_get_min_deflection(void);
+double glue_get_max_deflection(void);
+bool glue_validate_external_command(double deflection, double rate);
 
 #endif
